backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
bcrypt==4.1.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
textblob==0.17.1
librosa==0.10.1
soundfile==0.12.1
Pillow==10.1.0
python-dateutil==2.8.2
